refactor(interpreter): split activation register coordination
